Update Farbod_sclib.py
Fixed typos, 3 is not used in codes as an int and max_supplier variable handles it, behavioral functions were moved to the new Agents class as was suggested, but agent.all_agents is still in Agent class. Agents profits are now being calculated and added to their working capital. For suppliers and manufacturers this task is carried out in the delivery method and for retailers a new method was defined.